refactor orchestrator package
