+++ functions-python/tasks_executor/src/tasks/seal_of_reliability/orchestrator/seal_orchestrator.py
